% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessPack.R
\docType{data}
\name{chess.small}
\alias{chess.small}
\title{Data set and reference - Chess Game Dataset (Lichess)}
\format{
a data frame with 20019 observations. Raw data games.csv contains 16 columns, cleaned data chess.small contains 7 columns:
\describe{
 \item{rated}{(T/F)rated and unrated games}
 \item{turns}{number of moves in the game}
 \item{winner}{who won the game 'white', 'black'}
 \item{opening_eco}{code for opening move sets}
 \item{opening_ply} {Number of moves in the opening phase}
 \item{winner2}{dummy variable for wiinner 1=white, 0=black}
 \item{game_type}{Fischer-delay or sudden death}
 }

 @source \url{https://www.kaggle.com/datasets/datasnaek/chess}?
}
\usage{
chess.small
}
\description{
This is a set of just over 20,000 games collected from a selection of users
 on the site Lichess.org. This is a list of all columns in the raw data frame.
}
\keyword{datasets}
